I added header and footer
I added header and footer.
In the header I made all the links and also added clickable search and cart icons.
The footer is almost ready, I need to finish the screen responsiveness. I also removed the "Benefits" block from the footer, as I still think the best place for it is in the shop section.