Added landing page body
The landing page is complete with the carousel, shop and blog buttons (including animation), and "about us" section.
The layout is responsive and should look good on all devices.
However in this version, the header breaks the mobile layout. The body's responsiveness works fine on its own.
Further tweaks may be needed